removed duplicates colerated with translation feature
