+++ app/models/question.rb
@@ -8,6 +8,7 @@ class Question < ApplicationRecord
   include FormulaInterface
   include BlockHelper
   include Translate
+  include ::TranslationAuxiliaryMethods
 
   CURRENT_VERSION = '2'
 
@@ -125,17 +126,11 @@ class Question < ApplicationRecord
   end
 
   def translate_title(translator, source_language_name_short, destination_language_name_short)
-    original_text['title'] = title
-    new_title = translator.translate(title, source_language_name_short, destination_language_name_short)
-
-    update!(title: new_title)
+    translate_attribute('title', title, translator, source_language_name_short, destination_language_name_short)
   end
 
   def translate_subtitle(translator, source_language_name_short, destination_language_name_short)
-    original_text['subtitle'] = subtitle
-    new_subtitle = translator.translate(subtitle, source_language_name_short, destination_language_name_short)
-
-    update!(subtitle: new_subtitle)
+    translate_attribute('subtitle', subtitle, translator, source_language_name_short, destination_language_name_short)
   end
 
   def translate_image_description(translator, source_language_name_short, destination_language_name_short)
