+++ app/models/report_template/section/variant.rb
@@ -3,6 +3,7 @@
 class ReportTemplate::Section::Variant < ApplicationRecord
   has_paper_trail
   extend DefaultValues
+  include ::TranslationAuxiliaryMethods
 
   CURRENT_VERSION = '1'
 
@@ -21,17 +22,11 @@ class ReportTemplate::Section::Variant < ApplicationRecord
   ].freeze
 
   def translate_title(translator, source_language_name_short, destination_language_name_short)
-    original_text['title'] = title
-    new_title = translator.translate(title, source_language_name_short, destination_language_name_short)
-
-    update!(title: new_title)
+    translate_attribute('title', title, translator, source_language_name_short, destination_language_name_short)
   end
 
   def translate_content(translator, source_language_name_short, destination_language_name_short)
-    original_text['content'] = content
-    new_content = translator.translate(content, source_language_name_short, destination_language_name_short)
-
-    update!(content: new_content)
+    translate_attribute('content', content, translator, source_language_name_short, destination_language_name_short)
   end
 
   def translate(translator, source_language_name_short, destination_language_name_short)
