+++ app/models/session.rb
@@ -7,6 +7,7 @@ class Session < ApplicationRecord
   include FormulaInterface
   include InvitationInterface
   include Translate
+  include ::TranslationAuxiliaryMethods
 
   CURRENT_VERSION = '1'
 
@@ -146,10 +147,7 @@ class Session < ApplicationRecord
   end
 
   def translate_name(translator, source_language_name_short, destination_language_name_short)
-    original_text['name'] = name
-    new_name = translator.translate(name, source_language_name_short, destination_language_name_short)
-
-    update!(name: new_name)
+    translate_attribute('name', name, translator, source_language_name_short, destination_language_name_short)
   end
 
   def translate_sms_plans(translator, source_language_name_short, destination_language_name_short)
