+++ app/models/sms_plan/variant.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 class SmsPlan::Variant < ApplicationRecord
+  include ::TranslationAuxiliaryMethods
   has_paper_trail
   belongs_to :sms_plan
   has_one_attached :attachment, dependent: :purge_later
@@ -16,10 +17,7 @@ class SmsPlan::Variant < ApplicationRecord
   ATTR_NAMES_TO_COPY = %w[formula_match content].freeze
 
   def translate(translator, src_language_name_short, dest_language_name_short)
-    original_text['content'] = content
-    new_content = translator.translate(content, src_language_name_short, dest_language_name_short)
-
-    update!(content: new_content)
+    translate_attribute('content', content, translator, src_language_name_short, dest_language_name_short)
   end
 
   private
